+++ src/Kriti.hs
@@ -1,4 +1,4 @@
-module Kriti (RenderedError (..), ErrorCode (..), AlexSourcePos (..), ValueExt (..), runKriti) where
+module Kriti (SerializedError (..), ErrorCode (..), AlexSourcePos (..), ValueExt (..), runKriti) where
 
 import qualified Data.Aeson as J
 import Data.Bifunctor (first)
@@ -8,7 +8,7 @@ import Kriti.Error
 import Kriti.Eval
 import Kriti.Parser
 
-runKriti :: B.ByteString -> [(T.Text, J.Value)] -> Either RenderedError J.Value
+runKriti :: B.ByteString -> [(T.Text, J.Value)] -> Either SerializedError J.Value
 runKriti template source = do
-  template' <- first render $ parser template
-  first render $ runEval template' source
+  template' <- first serialize $ parser template
+  first serialize $ runEval template' source