+++ src/Kriti/Eval.hs
@@ -18,31 +18,31 @@ import qualified Network.URI as URI
 
 data EvalError
   = InvalidPath Span (V.Vector Accessor)
-  | TypeError Span T.Text
+  | TypeError ValueExt T.Text 
   | RangeError Span
-  deriving (Show)
+  deriving Show
 
-instance RenderError EvalError where
-  render :: EvalError -> RenderedError
-  render (InvalidPath span' path) = RenderedError {_code = InvalidPathCode, _message = "\"" <> renderVect path <> "\"", _span = span'}
-  render (TypeError span' txt) = RenderedError {_code = TypeErrorCode, _message = txt, _span = span'}
-  render (RangeError span') = RenderedError {_code = RangeErrorCode, _message = "Can only range over an array", _span = span'}
+instance SerializeError EvalError where
+  serialize :: EvalError -> SerializedError 
+  serialize (InvalidPath span' path) = SerializedError  {_code = InvalidPathCode, _message = "\"" <> renderVect path <> "\"", _span = span'}
+  serialize (TypeError term msg) = SerializedError  {_code = TypeErrorCode, _message = msg, _span = locate term}
+  serialize (RangeError span') = SerializedError  {_code = RangeErrorCode, _message = "Can only range over an array", _span = span'}
 
 type Ctxt = Compat.Object J.Value
 
 getSourcePos :: EvalError -> Span
 getSourcePos (InvalidPath pos _) = pos
-getSourcePos (TypeError pos _) = pos
+getSourcePos (TypeError term _) = locate term
 getSourcePos (RangeError pos) = pos
 
-evalPath :: J.Value -> V.Vector (Accessor) -> ExceptT EvalError (Reader Ctxt) J.Value
-evalPath ctx path =
+evalPath :: ValueExt -> J.Value -> V.Vector (Accessor) -> ExceptT EvalError (Reader Ctxt) J.Value
+evalPath term ctx path =
   let step :: Monad m => J.Value -> Accessor -> ExceptT EvalError m J.Value
       step (J.Object o) (Obj sp k _) = maybe (throwError $ InvalidPath sp path) pure $ Compat.lookup k o
       step (J.Array xs) (Arr sp i) = maybe (throwError $ InvalidPath sp path) pure $ xs V.!? i
       -- TODO: Should we extend this error message with the local Context?
-      step _ (Obj sp _ _) = throwError $ TypeError sp "Expected object"
-      step _ (Arr sp _) = throwError $ TypeError sp "Expected array"
+      step _ (Obj _ _ _) = throwError $ TypeError term "Expected object"
+      step _ (Arr _ _) = throwError $ TypeError term "Expected array"
    in foldlM step ctx path
 
 isString :: J.Value -> Bool
@@ -63,30 +63,30 @@ typoOfJSON J.Bool {} = "Boolean"
 typoOfJSON J.Null = "Null"
 
 eval :: ValueExt -> ExceptT EvalError (Reader Ctxt) J.Value
-eval = \case
+eval term = case term of
   String _ str -> pure $ J.String str
   Number _ i -> pure $ J.Number i
   Boolean _ p -> pure $ J.Bool p
   Null _ -> pure J.Null
   Object _ fields -> J.Object <$> traverse eval fields
-  StringTem sp ts -> do
+  StringTem _ ts -> do
     vals <- traverse eval ts
     str <-
       vals & flip foldlM mempty \acc -> \case
         J.String val' -> pure $ acc <> val'
         J.Number i -> pure $ acc <> TE.decodeUtf8 (BL.toStrict $ J.encode i)
         J.Bool p -> pure $ acc <> TE.decodeUtf8 (BL.toStrict $ J.encode p)
-        t -> throwError $ TypeError sp $ "Cannot interpolate type: '" <> typoOfJSON t <> "'."
+        t -> throwError $ TypeError term $ "Cannot interpolate type: '" <> typoOfJSON t <> "'."
     pure $ J.String str
   Array _ xs -> J.Array <$> traverse eval xs
   Path _ path -> do
     ctx <- ask
-    evalPath (J.Object ctx) path
-  Iff sp p t1 t2 ->
+    evalPath term (J.Object ctx) path
+  Iff _ p t1 t2 ->
     eval p >>= \case
       J.Bool True -> eval t1
       J.Bool False -> eval t2
-      p' -> throwError $ TypeError sp $ renderBL $ "'" <> J.encode p' <> "' is not a boolean."
+      p' -> throwError $ TypeError term $ renderBL $ "'" <> J.encode p' <> "' is not a boolean."
   Eq _ t1 t2 -> do
     res <- (==) <$> eval t1 <*> eval t2
     pure $ J.Bool res
@@ -98,39 +98,39 @@ eval = \case
     t1' <- eval t1
     t2' <- eval t2
     pure $ J.Bool $ t1' > t2'
-  And sp t1 t2 -> do
+  And _ t1 t2 -> do
     t1' <- eval t1
     t2' <- eval t2
     case (t1', t2') of
       (J.Bool p, J.Bool q) -> pure $ J.Bool $ p && q
-      (t1'', J.Bool _) -> throwError $ TypeError sp $ renderBL $ "'" <> J.encode t1'' <> "' is not a boolean."
-      (_, t2'') -> throwError $ TypeError sp $ renderBL $ "'" <> J.encode t2'' <> "' is not a boolean."
-  Or sp t1 t2 -> do
+      (t1'', J.Bool _) -> throwError $ TypeError term $ renderBL $ "'" <> J.encode t1'' <> "' is not a boolean."
+      (_, t2'') -> throwError $ TypeError term $ renderBL $ "'" <> J.encode t2'' <> "' is not a boolean."
+  Or _ t1 t2 -> do
     t1' <- eval t1
     t2' <- eval t2
     case (t1', t2') of
       (J.Bool p, J.Bool q) -> pure $ J.Bool $ p || q
-      (t1'', J.Bool _) -> throwError $ TypeError sp $ renderBL $ "'" <> J.encode t1'' <> "' is not a boolean."
-      (_, t2'') -> throwError $ TypeError sp $ renderBL $ "'" <> J.encode t2'' <> "' is not a boolean."
-  Member sp t ts -> do
+      (t1'', J.Bool _) -> throwError $ TypeError term $ renderBL $ "'" <> J.encode t1'' <> "' is not a boolean."
+      (_, t2'') -> throwError $ TypeError term $ renderBL $ "'" <> J.encode t2'' <> "' is not a boolean."
+  Member _ t ts -> do
     ts' <- eval ts
     case ts' of
       J.Array xs -> do
         t' <- eval t
         pure $ J.Bool $ t' `V.elem` xs
-      _ -> throwError $ TypeError sp $ T.pack $ "'" <> show ts' <> "' is not an array."
+      _ -> throwError $ TypeError term $ T.pack $ "'" <> show ts' <> "' is not an array."
   Range sp idx binder path body -> do
     ctx <- ask
-    pathResult <- evalPath (J.Object ctx) path
+    pathResult <- evalPath term (J.Object ctx) path 
     case pathResult of
       J.Array arr -> fmap J.Array . flip V.imapM arr $ \i val ->
         let newScope = [(binder, val)] <> [(idxBinder, J.Number $ fromIntegral i) | idxBinder <- maybeToList idx]
          in local (Compat.fromList newScope <>) (eval body)
       _ -> throwError $ RangeError sp
-  EscapeURI sp t1 -> do
+  EscapeURI _ t1 -> do
     t1' <- eval t1
     case t1' of
       J.String str ->
         let escapedUri = T.pack $ URI.escapeURIString URI.isUnreserved $ T.unpack str
          in pure $ J.String escapedUri
-      _ -> throwError $ TypeError sp $ T.pack $ "'" <> show t1' <> "' is not a string."
+      _ -> throwError $ TypeError term $ T.pack $ "'" <> show t1' <> "' is not a string."